Eugh, fix a crazy bug which resulted in walkhop times being read in incorrectly
Specifically, we were reading in floats, but writing (and expecting) doubles.
Add some asserts so this sort of problem will be easier to track down later.

+++ include/tripstop.h
@@ -44,7 +44,7 @@ struct WalkHop
     }
 
     int32_t dest_id;
-    double walktime;
+    float walktime;
 };
